requirements.txt
altgraph==0.17.2
dataclasses-json==0.5.5
flake8==3.9.2
marshmallow==3.13.0
marshmallow-enum==1.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
pycodestyle==2.7.0
pyflakes==2.3.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyYAML==5.4.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
